tools/requirements.txt
supabase>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
lxml>=4.9.0
